Bugfix and improvement in `common.sanitize`
At the top, it has to be `data.isImmutableJson` (which is either True
or False) instead of `t.isImmutableJson` (which is an attribute of the
ImmutableDict or ImmutableList type and always evaluates as True).

Also, instead of testing for specific types, this commit changes to
implementation to detect any mapping or sequence type (using
`collections.abc`).