Get rid of StateType, include raw and unresolved state under /sys
The concept of having three different "flavours" of state is awkward to
work with, because for any given path, you also need to specify whether
you are referencing the "raw" state (as loaded from the filesystem), the
"unresolved" state (with plugin states mounted, but symlinks not yet
resolved) or the "resolved" state (with symlinks resolved). Since our
paradigm is to represent the state of the whole system in one big tree,
it should be *one* tree, not three. So now there is only one state tree,
which is the resolved state. The "raw" state is now accessible under
"/sys/raw" and the unresolved state is in "/sys/unresolved".